requirements.txt
anyio==3.6.2
attrs==22.1.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
ecdsa==0.18.0
exceptiongroup==1.0.0
fastapi==0.85.1
h11==0.14.0
idna==3.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.19.0
